[CFIInstrInserter] Allow more general CSR locations.
Currently there are implicit assumptions that:
(1) each CSR can only be saved in one location,
(2) each location is either register or at a fixed offset from CFA

One example when (2) is not true is one spills of scalable vectors are
present. Assumption (1) may be broken if we enable handling of CSRs in
the register allocator.