[CFIInstrInserter] Tests for scenarios not currently handled by `CFIInstrInserter`.
We add an option to enable `CFIInstrInserter` on RISCV (`false` by
default). On RISCV there exists a realistic scenario (involing scalable
vectors) where we need CFIInstrInserter to handle `cfi_escape`. We add a
test for that. Also, we add a test for the scenario when a callee-saved
register is saved at different locations.